Move to new module name, removing warning.
