+++ .gitignore
@@ -159,12 +159,10 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
-src/bbprep/_version.py
 src/bbprepared/_version.py
 
 /dist
 /build
-src/bbprep.egg-info
 src/bbprepared.egg-info
 /.coverage
