--- src/bbprep/__init__.py
@@ -1,54 +1,0 @@
-"""bbprep package."""
-
-import warnings
-
-from bbprep import generators, selectors
-from bbprep._internal.ensemble.calculators import EnergyCalculator, Optimiser
-from bbprep._internal.ensemble.ensemble import Conformer, Ensemble
-from bbprep._internal.modifiers.distanced_functional_groups import (
-    ClosestFGs,
-    FurthestFGs,
-)
-from bbprep._internal.modifiers.modifier import Modifier
-from bbprep._internal.modifiers.random_functional_groups import RandomFGs
-from bbprep._internal.modifiers.reorient_panel import (
-    PanelBuildingBlock,
-    ReorientC1Panel,
-    ReorientC2Panel,
-    ReorientPanel,
-)
-from bbprep._internal.processes.angle import MinimiseAngle
-from bbprep._internal.processes.ditopicfitter import DitopicFitter
-from bbprep._internal.processes.planarfy import Planarfy
-from bbprep._internal.processes.process import Process, TargetProcess
-from bbprep._internal.processes.torsion import TargetTorsion
-
-warnings.warn(
-    "using `import bbprep` will be deprecated on 16th of June 2026. "
-    "Please use `import bbprepared`",
-    DeprecationWarning,
-    stacklevel=2,
-)
-
-__all__ = [
-    "ClosestFGs",
-    "Conformer",
-    "DitopicFitter",
-    "EnergyCalculator",
-    "Ensemble",
-    "FurthestFGs",
-    "MinimiseAngle",
-    "Modifier",
-    "Optimiser",
-    "PanelBuildingBlock",
-    "Planarfy",
-    "Process",
-    "RandomFGs",
-    "ReorientC1Panel",
-    "ReorientC2Panel",
-    "ReorientPanel",
-    "TargetProcess",
-    "TargetTorsion",
-    "generators",
-    "selectors",
-]
