--- src/bbprep/generators.py
@@ -1,25 +1,0 @@
-"""generators package."""
-
-from bbprep._internal.generators.etkdg import ETKDG
-from bbprep._internal.generators.generator import Generator
-from bbprep._internal.generators.geometry_scanner import GeometryScanner
-from bbprep._internal.generators.scanner_by_selector import (
-    SelectorDistanceScanner,
-)
-from bbprep._internal.generators.targets import (
-    AngleRange,
-    BondRange,
-    TorsionRange,
-)
-from bbprep._internal.generators.torsion_scanner import TorsionScanner
-
-__all__ = [
-    "ETKDG",
-    "AngleRange",
-    "BondRange",
-    "Generator",
-    "GeometryScanner",
-    "SelectorDistanceScanner",
-    "TorsionRange",
-    "TorsionScanner",
-]
