--- src/bbprep/selectors.py
@@ -1,26 +1,0 @@
-"""selectors package."""
-
-from bbprep._internal.selectors.all_selector import (
-    AllNonHSelector,
-    AllSelector,
-)
-from bbprep._internal.selectors.binders import BindersSelector
-from bbprep._internal.selectors.by_id import ByIdSelector
-from bbprep._internal.selectors.by_smarts import BySmartsSelector
-from bbprep._internal.selectors.deleters import DeletersSelector
-from bbprep._internal.selectors.notplacers import NotPlacersSelector
-from bbprep._internal.selectors.selector import NullSelector, Selector
-from bbprep._internal.selectors.xcomx import XCOMXSelector
-
-__all__ = [
-    "AllNonHSelector",
-    "AllSelector",
-    "BindersSelector",
-    "ByIdSelector",
-    "BySmartsSelector",
-    "DeletersSelector",
-    "NotPlacersSelector",
-    "NullSelector",
-    "Selector",
-    "XCOMXSelector",
-]
