+++ src/bbprepared/__init__.py
@@ -1,33 +1,34 @@
-"""New bbprepared module."""
+"""bbprepared module."""
 
-from bbprep import (
-    ClosestFGs,
-    Conformer,
-    DitopicFitter,
+from bbprepared import generators, selectors
+from bbprepared._internal.ensemble.calculators import (
     EnergyCalculator,
-    Ensemble,
-    FurthestFGs,
-    MinimiseAngle,
-    Modifier,
     Optimiser,
+)
+from bbprepared._internal.ensemble.ensemble import Conformer, Ensemble
+from bbprepared._internal.modifiers.distanced_functional_groups import (
+    ClosestFGs,
+    FurthestFGs,
+)
+from bbprepared._internal.modifiers.modifier import Modifier
+from bbprepared._internal.modifiers.random_functional_groups import RandomFGs
+from bbprepared._internal.modifiers.reorient_panel import (
     PanelBuildingBlock,
-    Planarfy,
-    Process,
-    RandomFGs,
     ReorientC1Panel,
     ReorientC2Panel,
     ReorientPanel,
-    TargetProcess,
-    TargetTorsion,
 )
-from bbprepared import generators, selectors
+from bbprepared._internal.processes.angle import MinimiseAngle
+from bbprepared._internal.processes.ditopicfitter import DitopicFitter
+from bbprepared._internal.processes.planarfy import Planarfy
+from bbprepared._internal.processes.process import Process, TargetProcess
+from bbprepared._internal.processes.torsion import TargetTorsion
 
 __all__ = [
     "ClosestFGs",
     "Conformer",
     "DitopicFitter",
     "EnergyCalculator",
-    "Ensemble",
     "Ensemble",
     "FurthestFGs",
     "MinimiseAngle",
