+++ src/bbprepared/_internal/processes/torsion.py
@@ -1,7 +1,7 @@
 import stk
 import stko
 
-from bbprep._internal.ensemble.ensemble import Conformer
+from bbprepared._internal.ensemble.ensemble import Conformer
 
 from .process import TargetProcess
