+++ src/bbprepared/generators.py
@@ -1,15 +1,17 @@
 """generators package."""
 
-from bbprep.generators import (
-    ETKDG,
+from bbprepared._internal.generators.etkdg import ETKDG
+from bbprepared._internal.generators.generator import Generator
+from bbprepared._internal.generators.geometry_scanner import GeometryScanner
+from bbprepared._internal.generators.scanner_by_selector import (
+    SelectorDistanceScanner,
+)
+from bbprepared._internal.generators.targets import (
     AngleRange,
     BondRange,
-    Generator,
-    GeometryScanner,
-    SelectorDistanceScanner,
     TorsionRange,
-    TorsionScanner,
 )
+from bbprepared._internal.generators.torsion_scanner import TorsionScanner
 
 __all__ = [
     "ETKDG",
