+++ src/bbprepared/selectors.py
@@ -1,17 +1,16 @@
 """selectors package."""
 
-from bbprep.selectors import (
+from bbprepared._internal.selectors.all_selector import (
     AllNonHSelector,
     AllSelector,
-    BindersSelector,
-    ByIdSelector,
-    BySmartsSelector,
-    DeletersSelector,
-    NotPlacersSelector,
-    NullSelector,
-    Selector,
-    XCOMXSelector,
 )
+from bbprepared._internal.selectors.binders import BindersSelector
+from bbprepared._internal.selectors.by_id import ByIdSelector
+from bbprepared._internal.selectors.by_smarts import BySmartsSelector
+from bbprepared._internal.selectors.deleters import DeletersSelector
+from bbprepared._internal.selectors.notplacers import NotPlacersSelector
+from bbprepared._internal.selectors.selector import NullSelector, Selector
+from bbprepared._internal.selectors.xcomx import XCOMXSelector
 
 __all__ = [
     "AllNonHSelector",
